--- src/test/java/eggventory/commands/help/Help.txt
@@ -1,44 +1,0 @@
-This displays the lists of commands that are defined.
-Type 'help <name>' to find out more about the command <name>. This only works for add, delete and edit.
-A star (*) next to the name means that the command is not enabled.
-
-add
-    add stock <Stock Code> <Quantity> <Description>
-    add stocktype <Stock Type>
-
-delete
-    delete stock <Stock Code>
-    delete stocktype <Stock Type>
-    delete template <Template Name>
-    delete person <Matric No.>
-
-edit
-    edit stock <Property> <New Value>
-    edit person <Property> <New Value>
-
-*list
-    list stock
-    list stocktypes
-    list stocktype <Stock Type>
-    list loan
-    list template
-    list lost
-
-*find
-    find stock <Query>
-    find stocktype <Query>
-
-*loan
-    loan add <Matric No.> {<Stock Code> <Quantity>}
-    OR
-    loan add <Matric No.> <Template Name>
-    loan returned <Matric No.> {<Stock Code> <Quantity>}
-    OR
-    loan returned <Matric No.> <Template Name>
-
-*lost
-    lost <Stock Code> <Quantity>
-*undo
-    undo
-*redo
-    redo
