--- src/test/java/eggventory/commands/help/HelpAdd.txt
@@ -1,15 +1,0 @@
-Add new items or item types into the database.
-
-add:    add stock <Stock Type> <Stock Code> <Quantity> <Description>
-        add stocktype <Stock Type>
-
-<Stock Type>        The category name of the stock
-<Stock Code>        The model number of the new stock
-<Quantity>          The initial number of stock to be added into Eggventory.
-                    If you intend to increase the quantity of the stock, please use 'edit'
-<Description>       A short description of the stock
-<Stock Type>        The category name of the stock
-
-E.g. add stock resistor 12345 50 100ohm
-E.g. add stock resistor 23456 70 800ohm
-E.g. add stocktype resistor
