--- src/test/java/eggventory/commands/help/HelpDelete.txt
@@ -1,9 +1,0 @@
-Deletes existing item or item type in the database.
-
-delete:     delete stock <Stock Code>
-            delete stocktype <Stock Type>
-
-<Stock Code>        The model number of the existing stock
-<Stock Type>        The category name of the stock
-
-You can find out existing stock code and stock type by using 'list'
