--- src/test/java/eggventory/commands/help/HelpEdit.txt
@@ -1,9 +1,0 @@
-Directly modifies the value of a property of a stock.
-
-edit:       edit stock (<Stock Type>| <Stock Code> | <Quantity> | <Description> ) <New Value>
-
-<Stock Type>        The category name of the stock
-<Stock Code>        The model number of the existing stock
-<Quantity>          The number of stock in the inventory
-<Description>       The description of the stock
-<New Value>         The value of the property mentioned will be replaced by this new value
