--- src/test/java/eggventory/commands/help/HelpFind.txt
@@ -1,1 +1,0 @@
-helpfind
