--- src/test/java/eggventory/commands/help/HelpList.txt
@@ -1,1 +1,0 @@
-helplist
