--- src/test/java/eggventory/commands/help/HelpLoan.txt
@@ -1,1 +1,0 @@
-helploan
