--- src/test/java/eggventory/commands/help/HelpLost.txt
@@ -1,1 +1,0 @@
-helplost.
