--- src/test/java/eggventory/commands/help/HelpUndo.txt
@@ -1,1 +1,0 @@
-undo
