+++ src/test/java/eggventory/logic/HelpCommandTest.java
@@ -1,0 +1,7 @@
+package eggventory.logic;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+class HelpCommandTest {
+
+}
