+++ src/test/java/eggventory/logic/commands/ByeCommandTest.java
@@ -1,10 +1,10 @@
-package eggventory.commands.edit;
+package eggventory.logic.commands;
 
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class EditStockCommandTest {
+class ByeCommandTest {
 
     @Test
     void execute() {
