+++ src/test/java/eggventory/logic/commands/edit/EditStockCommandTest.java
@@ -1,10 +1,10 @@
-package eggventory.commands;
+package eggventory.logic.commands.edit;
 
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class ByeCommandTest {
+class EditStockCommandTest {
 
     @Test
     void execute() {
